Fix date handling logic in create_sprint_boards (#12)
* Fix date handling logic in create_sprint_boards

The loop would break prematurely because it was breaking when
today > sprint start, which would be the case for all sprints that
started in the past. Instead, the loop should break when that case
is not true anymore, and use the last matching line.

This also reenables the command in the board maintenance action.